add surrogates.py
